Data Management added!
-Save & Load generally work now

Main Problem:
* Saving several times has not been handled

+++ src/game_object/enemy/Alien.java
@@ -12,11 +12,11 @@ public class Alien extends Enemy
 
     /**
      *  Constructs the modern enemy for period 1
-     *  @param x - x coordinate
+     * @param x - x coordinate
      *  @param y - y coordinate
-     *  @param id - id of the object
+     * @param id - id of the object
      */
-    public Alien(float x, float y, ObjectID id )
+    public Alien(double x, double y, ObjectID id )
     {
         super(x, y, id);
