+++ src/game_object/player/Robot.java
@@ -27,12 +27,11 @@ public class Robot extends Character
 
     /**
      * Constructing the character with given parameters.
-     *
-     * @param x                 - x coordinate of the character.
+     *  @param x                 - x coordinate of the character.
      * @param y                 - y coordinate of the character.
      * @param id                - id of the character as a game object.
      */
-    public Robot (float x, float y, ObjectID id)
+    public Robot (double x, double y, ObjectID id)
     {
         super(x, y, id);
